Open source jammr server-side code
This is the server-side code for the jamming platform. See docs/ on how
to build and run it. Note that a secrets.yml file as well as individual
ssh key and TLS certificate files are also required to build and deploy.

This code powers jammr.net and was originally developed by jammr ltd. It
is now open source under the MIT license. Please note that its
dependencies have various licenses when considering the overall license.

If you run another instance of the jamming platform on your own domain,
please rebrand it to avoid confusion with jammr.net.

Signed-off-by: Stefan Hajnoczi <stefanha@jammr.net>